fix: fetch-http-handler returns body as null
When doing a http request that does not return a body, using the fetch api, in some browser's client versions, such chrome version 103.x.x, the body is returned as null in the response, and the deserializers always expect the body to have a value, even if this value is empty. What I do to fix this issue is to add a fallback for when the body is null or undefined we set a new Uint8Array() to the body.